requirements.txt
alpaca-trade-api
ccxt
numpy
pandas
pandas-market-calendars
paramiko
pillow
pydantic
pydantic-settings
python-dotenv
scikit-learn
yfinance

# Optional ML backend for models/ml_model.py
tensorflow
